requirements.txt
cdsapi
pyyaml
pandas
xarray
cfgrib